+++ src/dropin-model.js
@@ -70,7 +70,6 @@ DropinModel.prototype.asyncDependencyFailed = function (options) {
 
 DropinModel.prototype._checkAsyncDependencyFinished = function () {
   if (this.dependenciesInitializing === 0) {
-    this._emit('loadEnd');
     this._emit('asyncDependenciesReady');
   }
 };