+++ src/dropin.js
@@ -23,6 +23,7 @@ function Dropin(options) {
   this._dropinWrapper.id = 'braintree--dropin__' + this._componentID;
   this._dropinWrapper.setAttribute('data-braintree-id', 'wrapper');
   this._dropinWrapper.style.display = 'none';
+  this._dropinWrapper.className = 'braintree-loading';
   this._merchantConfiguration = options.merchantConfiguration;
 
   EventEmitter.call(this);