+++ CHANGELOG.md
@@ -2,12 +2,13 @@ CHANGELOG
 =========
 
 unreleased
------------
+----------
 - Use PayPal Checkout for PayPal View
 - Use version 3.12.0 of braintree-web
+- Improve loading transition
 
 1.0.0-beta.6
--------------
+------------
 - Disable payment methods if they error when creating
 - Add `paymentOptionPriority` option for specifying the ordering of payment options such as `card` and `paypal`
 - Add translations